# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Kullanıcı girişi",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Başarılı giriş"
          },
          "401": {
            "description": "Geçersiz kimlik bilgileri"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register",
        "summary": "Yeni kullanıcı kaydı",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kullanıcı başarıyla oluşturuldu"
          },
          "400": {
            "description": "Geçersiz veri"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "getProfile",
        "summary": "Kullanıcı profili görüntüleme",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Kullanıcı profili"
          },
          "401": {
            "description": "Yetkisiz erişim"
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/bookings": {
      "post": {
        "operationId": "create",
        "summary": "Create new booking",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created successfully"
          }
        },
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "findAll",
        "summary": "Get all bookings",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all bookings"
          }
        },
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/bookings/{id}": {
      "get": {
        "operationId": "findOne",
        "summary": "Get booking details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return booking details"
          }
        },
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "update",
        "summary": "Update booking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking updated successfully"
          }
        },
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "remove",
        "summary": "Delete booking (Admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking deleted successfully"
          }
        },
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/bookings/{id}/cancel": {
      "put": {
        "operationId": "cancel",
        "summary": "Cancel booking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking cancelled successfully"
          }
        },
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/blog": {
      "post": {
        "operationId": "create",
        "summary": "Create new blog post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog post created successfully"
          }
        },
        "tags": [
          "Blog"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "findAll",
        "summary": "Get all blog posts",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all blog posts"
          }
        },
        "tags": [
          "Blog"
        ]
      }
    },
    "/blog/search": {
      "get": {
        "operationId": "search",
        "summary": "Search blog posts",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return matching blog posts"
          }
        },
        "tags": [
          "Blog"
        ]
      }
    },
    "/blog/{id}": {
      "get": {
        "operationId": "findOne",
        "summary": "Get blog post by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return blog post details"
          }
        },
        "tags": [
          "Blog"
        ]
      },
      "put": {
        "operationId": "update",
        "summary": "Update blog post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blog post updated successfully"
          }
        },
        "tags": [
          "Blog"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "remove",
        "summary": "Delete blog post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog post deleted successfully"
          }
        },
        "tags": [
          "Blog"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/blog/by-slug/{slug}": {
      "get": {
        "operationId": "findBySlug",
        "summary": "Get blog post by slug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return blog post details"
          }
        },
        "tags": [
          "Blog"
        ]
      }
    },
    "/blog/{id}/like": {
      "put": {
        "operationId": "like",
        "summary": "Like blog post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog post liked successfully"
          }
        },
        "tags": [
          "Blog"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/pages/by-slug/{slug}": {
      "get": {
        "operationId": "findBySlug",
        "summary": "Get page by slug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return page details"
          }
        },
        "tags": [
          "Pages"
        ]
      }
    },
...
  },
  "info": {
    "title": "HizmetBul API",
    "description": "HizmetBul Servis Platformu API Dokümantasyonu",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "provider",
              "customer"
            ],
            "example": "customer"
          },
          "userType": {
            "type": "string",
            "enum": [
              "customer",
              "provider"
            ]
          }
        },
        "required": [
          "email",
          "password",
          "fullName",
          "role",
          "userType"
        ]
      },
...
      "LocationPropertiesDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "İstiklal Caddesi No: 50, Beyoğlu"
          },
          "city": {
            "type": "string",
            "example": "İstanbul"
          },
          "state": {
            "type": "string",
            "example": "İstanbul"
          }
        },
        "required": [
          "address",
          "city",
          "state"
        ]
      },
      "LocationDto": {
        "type": "object",
        "properties": {
          "lng": {
            "type": "number",
            "example": 28.9784
          },
          "lat": {
            "type": "number",
            "example": 41.0082
          },
          "properties": {
            "$ref": "#/components/schemas/LocationPropertiesDto"
          }
        },
        "required": [
          "lng",
          "lat",
          "properties"
        ]
      },
...
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "serviceId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-20"
          },
          "time": {
            "type": "string",
            "example": "14:30"
          },
          "location": {
            "$ref": "#/components/schemas/LocationDto"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "serviceId",
          "date",
          "time"
        ]
      },
      "UpdateBookingDto": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-20"
          },
          "time": {
            "type": "string",
            "example": "14:30"
          },
          "location": {
            "$ref": "#/components/schemas/LocationDto"
          },
          "notes": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "completed",
              "cancelled"
            ],
            "description": "Booking status"
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "pending",
              "paid",
              "refunded",
              "failed"
            ],
            "description": "Payment status"
          }
        }
      },
      "CancelBookingDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Schedule conflict",
            "description": "Reason for cancellation"
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "CreateBlogDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Blog Başlığı"
          },
          "slug": {
            "type": "string",
            "example": "blog-basligi"
          },
          "content": {
            "type": "string",
            "example": "Blog içeriği..."
          },
          "excerpt": {
            "type": "string",
            "example": "Kısa özet..."
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "featuredImage": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "seo": {
            "type": "object"
          }
        },
        "required": [
          "title",
          "slug",
          "content"
        ]
      },
      "UpdateBlogDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Blog Başlığı"
          },
          "slug": {
            "type": "string",
            "example": "blog-basligi"
          },
          "content": {
            "type": "string",
            "example": "Blog içeriği..."
          },
          "excerpt": {
            "type": "string",
            "example": "Kısa özet..."
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "featuredImage": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "seo": {
            "type": "object"
          }
        }
      },
...
    }
  }
}